sdn: add frr config generation helpers
Adds a new method to the SDN module that is responsible for generating
and writing the FRR configuration for all SDN plugins combined. It
utilizes the newly introduced FRR helper as well as the newly
introduced API for the controllers to generate an frr_config instead
of generating the configuration in the controller directly. It can
also reload the FRR daemon.

Change the tests to use this new API as well, so they use the new
methods for generating the frr configuration. They previously used a
different code-path for generating the FRR config compared to the
actual worker task, so this also ensures that tests validate the
configuration that *actually* gets generated.

Co-authored-by: Gabriel Goller <g.goller@proxmox.com>
Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20250716130837.585796-42-g.goller@proxmox.com