test: isis: add test for standalone configuration
With how the config generation worked before, it was not possible to
create a standalone isis controller, since the FRR config was only
generated with an existing EVPN controller. Since each controller is
now responsible for creating its own configuration, it is possible to
create a standalone isis controller without having any evpn controller
configured. Add a test that covers that scenario.

Co-authored-by: Gabriel Goller <g.goller@proxmox.com>
Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20250716130837.585796-44-g.goller@proxmox.com