frr: early exit on apply when frr is not installed
When frr isn't installed (happens e.g. on pve8 to 9 update) then the SDN
and Networking "Apply Configuration" task will throw an error because it
can't access the frr daemons file to check if the daemons are enabled.
In order to fix this, check if the frr config file exists before
continuing with the frr config generation.

Also make the "daemons file does not exist" error nicer.

Signed-off-by: Gabriel Goller <g.goller@proxmox.com>
Link: https://lore.proxmox.com/20250723154934.928349-1-g.goller@proxmox.com
 [TL: switch to checking just for /etc/frr and perltidy]
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>