api: add lock and release endpoints for global configuration lock
This endpoint exposes the newly introduced global lock functionality
via the API. It adds endpoints for acquiring and releasing the lock.
Acquiring the lock is as simple as:

  pvesh create /cluster/sdn/lock

The flag 'allow-pending' governs whether the lock should be acquired
if there are pending configuration changes.

The release endpoint can also be used to forcibly release the lock via
CLI without requiring the knowledge of the lock-token:

  pvesh delete /cluster/sdn/lock --force 1

Co-authored-by: Gabriel Goller <g.goller@proxmox.com>
Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20250729092933.90118-5-g.goller@proxmox.com