frr: print task warning when frr reload fails
Instead of just printing a normal warning, use log_warn, which adds a
warning to the whole task. This means the whole task ends with a
"warning" status which is visible nicely in the task list. Previously we
just printed a warning, which made the task end with "OK" even if the
reload failed.

frr-reload.py only fails if it's not installed (where a warning status
is ok) or the configuration is not valid (where a warning status is also
ok). Having frr-reload.py uninstalled for some reason and always relying
on `systemctl restart frr` to reload the sdn config is kinda weird and
you probably deserve to get a warning on each sdn apply task.

Signed-off-by: Gabriel Goller <g.goller@proxmox.com>
Link: https://lore.proxmox.com/20250911153359.278834-1-g.goller@proxmox.com