sdn: write frr config if the frr.conf.local file exists
Currently we check if there is a fabric or a controller existing to
decide if we write the frr.conf file. Also add a check if the
frr.conf.local exists. This way it's possible to have a custom config in
the frr.conf.local file without any SDN objects configured and still
be able to apply the config.
By not checking the content we can also clear the frr config by clearing
the frr.conf.local file.

Reported-by: Hannes Dürr <h.duerr@proxmox.com>
Signed-off-by: Gabriel Goller <g.goller@proxmox.com>
Reviewed-by: Stefan Hanreich <s.hanreich@proxmox.com>
Tested-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20251002130950.44736-1-g.goller@proxmox.com