refactor: rework api module structure for the /nodes/{node}/sdn subdir
Previously, the endpoints for the SDN browser were contained in the
PVE::API2::Network::SDN::Zones module, which was registered in
PVE::API2::Nodes (located in pve-manager) directly with the sub-path
'sdn/{zone}'.

Instead of importing the zones module in pve-manager directly, create
a top-level Nodes module in pve-network, that is then imported in
pve-manager under the path 'sdn'. By importing just the root-level
module, changes can be made to the pve-network API without having to
update pve-manager. This allows for a better module structure on
pve-network side, because different entities can then be managed in
pve-network directly in the new top-level module. This is mainly in
preparation for adding new Vnets and Fabrics modules.

Also, the zones' content endpoint was contained in its own module. In
order to make it easier to add new endpoints to the 'sdn/zones/{zone}'
directory, move the contents endpoint into a new module that is
imported directly under 'sdn/zones/{zone}'. This allows easily adding
endpoints under the 'sdn/zones/{zone}' sub directory, without having
to create a separate module for each endpoint and registering it in
the Status module.

No functional changes intended, every endpoint should be the same as
before (URL + returned information).

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20251113161926.793020-4-s.hanreich@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>